requirement.txt
langgraph
langchain-core
langchain_openai
duckduckgo-search
langchain-community
python-dotenv
fastapi
uvicorn
pytest
